reworked the codes in STL folder
