+++ Code/CMakeLists.txt
@@ -30,5 +30,6 @@ add_subdirectory( Threads )
 add_subdirectory( Traits )
 add_subdirectory( TypeDeduction )
 add_subdirectory( move_examples )
+add_subdirectory( STL )
 add_subdirectory( TypesAndValues )
 add_subdirectory( Alignment )
