+++ Code/STL/integer_sequence/CMakeLists.txt
@@ -1,0 +1,5 @@
+file( GLOB SOURCES RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} *.cpp )
+foreach( target_s ${SOURCES} )
+    string( REPLACE ".cpp" "" target_n ${target_s} )
+    add_executable( ${target_n} ${target_s} )
+endforeach()
