+++ Code/STL/tuple/CMakeLists.txt
@@ -1,0 +1,4 @@
+
+
+add_executable( tuple_main tuple_main.cpp )
+add_executable( tie tie.cpp)
