--- Code/STL/tuple/get.cpp
@@ -1,2 +1,0 @@
-template<int Id, typename Tuple>
-constexpr auto get(Tuple t_){ return t_.template get<Tuple::size-Id>();}
