--- Code/STL/tuple/m_get.cpp
@@ -1,2 +1,0 @@
-    template<int Id>
-    constexpr auto get() const {return Id==size ? m_value : super::template get<Id>();}
